Fix: data URL params missing page and sort
- Add: sort direction to the URL params
- Fix: page data is lost when hydrating from URL param when
       search or date input exists

The page data problem occured because of a circular reference.
When the page is loaded with a search or date query param, the
searchInput, dateFromInput, or dateToInput values were set.
This fired the watchers which reset the currentPage to 1.

This could not be handled by a simple flag in the watcher,
because these watchers were debounced.

This was solved by using the watchIgnorable() utility from
@vueuse.