Update story event position
This change replaces the left/center/right position of the story events
with a percentage-based position, so that the story event is positioned
over its hex based on where it is in the overall timeline.

This fixes an issue where sometimes a story would scroll backwards as
it moved from a left-positioned event to a center-positioned event. Now,
each new event in the story adds a rightward scroll.